@@ -1,13 +1,11 @@
 package com.consulta.repositorio;
 
-import java.util.Optional;
-
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
+import org.springframework.security.core.userdetails.User;
 
 import com.consulta.model.Usuario;
 
 public interface IUsuarioRepo extends JpaRepository<Usuario, Integer>{
 	
-  	
+  	public User findByUsername(String username);
 }
